MolecularIntegrals.jl

The goal of MolecularIntegrals.jl is to supply fast and hackable one- and two-electron integrals for computational chemistry calculations. There are many excellent molecular integral packages available to Julia programmers, but few are written in Julia. This project will explore how fast we can make these integrals while maintaining a readable and hackable code base.

One Electron Integrals

Two Electron Integrals

MolecularIntegrals.jl supports a slower method based on Huzinaga's work, and a faster set of integrals based on Head-Gordon and Pople's work [HGP].

Huzinaga Method

Head-Gordan/Pople Recurrance Relations

MolecularIntegrals.vrr_array — Method

vrr_array(amax,cmax, aexpn,bexpn,cexpn,dexpn, A,B,C,D)

Use Head-Gordon/Pople's vertical recurrence relations to compute an array of two-electron integrals.

A, B, C, D are the centers of four Gaussian functions. aexpn, bexpn, cexpn, dexpn are their exponents. amax and cmax are related to the sum of the shell angular momenta for the a+b, and c+d shells, respectively.

The function returns an nxm array, where n is the number of aos in the a+b shell, and m is the number of aos in the c+d shell.

The speeds of vrr_array and vrr_wide_array are roughly equivalent, and both interfaces are being retained for convenience.

Missing docstring.

Missing docstring for hrr_array(amax,cmax, aexpn,bexpn,cexpn,dexpn, A,B,C,D). Check Documenter's build log for details.

Other Methods

Rys quadrature

References

HGPA method for two-electron Gaussian integral and integral derivative evaluation using recurrence relations. Martin Head-Gordon and John A. Pople. JCP, 89 (9), 5777, 1988.

GMolecular Integrals Over Gaussian Basis Functions. Peter M. W. Gill. Adv. Q. Chem., 25, 141 (1994).

GPThe Prism Algorithm for Two-Electron Integrals. Peter M. W. Gill and John A. Pople. IJQC, 40, 753 (1991).