MolecularIntegrals.jl

The goal of MolecularIntegrals.jl is to supply fast and hackable one- and two-electron integrals for computational chemistry calculations. There are many excellent molecular integral packages available to Julia programmers, but few are written in Julia. This project will explore how fast we can make these integrals while maintaining a readable and hackable code base.

MolecularIntegrals.jl strives to leverage the excellent work done by similar projects:

libints and its Julia bindings Lints.jl

Pyscf, the libcint package.

JuliaChem.jl's JERI bindings

PyQuante's python and c/cython integrals, and the experimental Julia version

One Electron Integrals

MolecularIntegrals.overlap — Function

overlap(a::PGBF,b::PGBF)

Compute the overlap between primitive Gaussian basis functions a and b.

overlap(a::CGBF,b::CGBF)

Compute the overlap between contracted Gaussian basis functions a and b.

overlap(aexpn,ax,ay,az,aI,aJ,aK,bexpn,bx,by,bz,bI,bJ,bK)

Compute the overlap between primitive Gaussian basis functions defined by centers ax,ay,az, bx,by,bz, powers aI,aJ,aK bI,bJ,bK, and exponents aexpn,bexpn.

overlap(aexpn,axyz,aI,aJ,aK,bexpn,bxyz,bI,bJ,bK)

Compute the overlap between primitive Gaussian basis functions defined by centers axyz, bxyz, powers aI,aJ,aK bI,bJ,bK, and exponents aexpn,bexpn.

MolecularIntegrals.kinetic — Function

kinetic(a::PGBF,b::PGBF)

Compute the kinetic energy between primitive Gaussian basis functions a and b.

kinetic(aexpn,ax,ay,az,aI,aJ,aK,bexpn,bx,by,bz,bI,bJ,bK)

Compute the kinetic energy between primitive Gaussian basis functions defined by centers ax,ay,az, bx,by,bz, powers aI,aJ,aK bI,bJ,bK, and exponents aexpn,bexpn.

kinetic(aexpn,axyz,aI,aJ,aK,bexpn,bxyz,bI,bJ,bK)

Compute the kinetic energy between primitive Gaussian basis functions defined by centers ax,yz, bxyz, powers aI,aJ,aK bI,bJ,bK, and exponents aexpn,bexpn.

kinetic(a::CGBF,b::CGBF)

Compute the kinetic energy between contracted Gaussian basis functions a and b.

MolecularIntegrals.nuclear_attraction — Function

nuclear_attraction(a::PGBF,b::PGBF,cxyz)

Compute the nuclear attraction energy between primitive Gaussian basis functions a and b and center cxyz.

nuclear_attraction(a::PGBF,b::PGBF,c::Atom)

Compute the nuclear attraction energy between primitive Gaussian basis functions a and b and atom c.

nuclear_attraction(a::PGBF,b::PGBF,m::Vector{Atom})

Compute the sum of nuclear attraction energy between primitive Gaussian basis functions a and b and the vector of atoms in m.

nuclear_attraction(aexpn,ax,ay,az,aI,aJ,aK,bexpn,bx,by,bz,bI,bJ,bK,cx,cy,cz)

Compute the nuclear attraction energy between primitive Gaussian basis functions defined by ax,ay,az, bx,by,bz, powers aI,aJ,aK, bI,bJ,bK, and exponents aexpn,bexpn and center cxyz.

nuclear_attraction(aexpn,ax,ay,az,aI,aJ,aK,bexpn,bx,by,bz,bI,bJ,bK,cx,cy,cz)

Compute the nuclear attraction energy between primitive Gaussian basis functions defined by ax,ay,az, bx,by,bz, powers aI,aJ,aK, bI,bJ,bK, and exponents aexpn,bexpn and center cxyz.

nuclear_attraction(a::CGBF,b::CGBF,cxyz)

Compute the nuclear attraction energy between contracted Gaussian basis functions a and b and center cxyz.

nuclear_attraction(a::CGBF,b::CGBF,c::Atom)

Compute the nuclear attraction energy between contracted Gaussian basis functions a and b and atom c.

nuclear_attraction(a::CGBF,b::CGBF,m::Vector{Atom})

Compute the sum of nuclear attraction energy between contracted Gaussian basis functions a and b and vector of atoms in m.

Two Electron Integrals

MolecularIntegrals.jl supports a slower method based on Huzinaga's work, and a faster set of integrals based on Head-Gordon and Pople's work [HGP].

The interface to the Huzinaga integrals is

MolecularIntegrals.coulomb — Function

coulomb(aexpn,ax,ay,az,aI,aJ,aK,
    bexpn,bx,by,bz,bI,bJ,bK,
    cexpn,cx,cy,cz,cI,cJ,cK,
    dexpn,dx,dy,dz,dI,dJ,dK)

Compute the coulomb repulsion between four primitive Gaussian basis functions defined by their exponents, xyz coordinates, and IJK powers.

coulomb(a::PGBF,b::PGBF,c::PGBF,d::PGBF)

Compute the coulomb repulsion between four primitive Gaussian basis functions a,b,c,d.

coulomb(a::CGBF,b::CGBF,c::CGBF,d::CGBF)

Compute the coulomb repulsion between four contracted Gaussian basis functions a,b,c,d.

These are useful for checking the values of other integrals, and have a simple, mature interface.

Head-Gordon/Pople Recurrence Relations

Integrals are computed using Head-Gordon and Pople's[HGP] recurrence relations using vertical (VRR) and horizontal (HRR) recurrence relations. In the notation of [HGP], vertical recurrence relations construct integrals of the form [a0|c0] from kernels [00|00]^m via eq 6. Horizontal recurrence relations construct integrals of the form [ab|cd] from kernels [a0|c0].

Vertical recurrence relations

MolecularIntegrals.jl returns VRRs either as a 2d array using vrr_array, or as a 6d array using vrr_widearray:

MolecularIntegrals.vrr_array — Function

vrr_array(amax,cmax, aexpn,bexpn,cexpn,dexpn, A,B,C,D)

Use Head-Gordon/Pople's vertical recurrence relations to compute an array of two-electron integrals.

A, B, C, D are the centers of four Gaussian functions. aexpn, bexpn, cexpn, dexpn are their exponents. amax and cmax are related to the sum of the shell angular momenta for the a+b, and c+d shells, respectively.

The function returns an nxm array, where n is the number of aos in the a+b shell, and m is the number of aos in the c+d shell.

The speeds of vrr_array and vrr_widearray are roughly equivalent, and both interfaces are being retained for convenience.

MolecularIntegrals.vrr_widearray — Function

vrr_widearray(amax,cmax, aexpn,bexpn,cexpn,dexpn, A,B,C,D)

Use Head-Gordon/Pople's vertical recurrence relations to compute an array of integrals.

A, B, C, D are the centers of four Gaussian functions. aexpn, bexpn, cexpn, dexpn are their exponents. amax and cmax are related to the sum of the shell angular momenta for the a+b, and c+d shells, respectively.

The function returns a six-dimensional array over the possible powers of the a+b and c+d shell functions.

The speeds of vrr_array and vrr_widearray are roughly equivalent, and both interfaces are being retained for convenience.

Horizontal recurrence relations

MolecularIntegrals.hrr_array — Function

hrr_array(ashell,bshell,cshell,dshell, aexpn,bexpn,cexpn,dexpn, A,B,C,D)

Use Head-Gordon/Pople's horizontal recurrence relations to compute an array of two-electron integrals.

A, B, C, D are the centers of four Gaussian functions. aexpn, bexpn, cexpn, dexpn are their exponents. ashell, bshell, cshell and dshell are the shell angular momenta for the a, b, c, and d shells, respectively.

The function returns a (k,l,m,n)-dimensional array, where the dimensions correspond to the number of aos in the a,b,c,d shells.

MolecularIntegrals.hrr_dict — Function

hrr_dict(ashell,bshell,cshell,dshell, aexpn,bexpn,cexpn,dexpn, A,B,C,D)

Use Head-Gordon/Pople's horizontal recurrence relations to compute an array of two-electron integrals.

A, B, C, D are the centers of four Gaussian functions. aexpn, bexpn, cexpn, dexpn are their exponents. ashell, bshell, cshell and dshell are the shell angular momenta for the a, b, c, and d shells, respectively.

The function returns a dictionary containing entries for the relevant integrals. E.g., hrrs[ax,ay,az,bx,by,bz,cx,cy,cz,dpx,dpy,dpz] contains the integral corresponding to the bfs with powers ax,ay,az, bx,by,bz, cx,cy,cz, dx,dy,dz.

hrr_dict is slower than hrr_array, but is kept for convenience.

References

HGPA method for two-electron Gaussian integral and integral derivative evaluation using recurrence relations. Martin Head-Gordon and John A. Pople. JCP, 89 (9), 5777, 1988.